change parent info and docker.prefix
