+++ admin-service/src/main/java/dev/cfernandezcairo/microservices/AdminApplication.java
@@ -1,0 +1,19 @@
+package com.ev.microservices;
+
+import de.codecentric.boot.admin.server.config.EnableAdminServer;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+import org.springframework.context.annotation.Configuration;
+
+@SpringBootApplication
+@Configuration
+@EnableAutoConfiguration
+@EnableAdminServer
+@EnableDiscoveryClient
+public class AdminApplication {
+    public static void main(String[] args){
+        SpringApplication.run(AdminApplication.class, args);
+    }
+}
