--- config-service/src/main/java/com/ev/microservices/ConfigApplication.java
@@ -1,15 +1,0 @@
-package com.ev.microservices;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.cloud.config.server.EnableConfigServer;
-
-@SpringBootApplication
-@EnableConfigServer
-@EnableDiscoveryClient
-public class ConfigApplication {
-    public static void main(String[] args){
-        SpringApplication.run(ConfigApplication.class, args);
-    }
-}
