+++ eureka-service/src/main/java/dev/cfernandezcairo/microservices/EurekaApplication.java
@@ -1,0 +1,13 @@
+package com.ev.microservices;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+@SpringBootApplication
+@EnableEurekaServer
+public class EurekaApplication {
+    public static void main(String[] args){
+        SpringApplication.run(EurekaApplication.class, args);
+    }
+}
