+++ gateway-service/src/main/java/dev/cfernandezcairo/microservices/GatewayApplication.java
@@ -1,0 +1,13 @@
+package com.ev.microservices;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class GatewayApplication {
+    public static void main(String[] args){
+        SpringApplication.run(GatewayApplication.class, args);
+    }
+}
