+++ pom.xml
@@ -9,9 +9,9 @@
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
 
-    <groupId>com.ev.microservices</groupId>
-    <artifactId>microservice-base</artifactId>
-    <version>1.0.0</version>
+    <groupId>dev.cfernandezcairo</groupId>
+    <artifactId>wastemanager-microservice-base</artifactId>
+    <version>1.0-SNAPSHOT</version>
 
     <packaging>pom</packaging>
     <name>microservice-base</name>
@@ -27,7 +27,7 @@
 
         <docker.image.prefix>microservices/base</docker.image.prefix>
         <docker-maven-plugin.version>1.2.2</docker-maven-plugin.version>
-        <docker.image.prefix>ectricalvehicle/base</docker.image.prefix>
+        <docker.image.prefix>april2024/base</docker.image.prefix>
     </properties>
 
     <url></url>
